Jeffreys confidence intervals for precision/recall
Features:
- Adds modified Jeffrey's confidence intervals for precision/recall
  (Brown, Cai and DasGupta: Interval Estimation for a Binomial Proportion.
  2001, doi:10.1214/ss/1009213286)
- Adds and updates test cases for som.py
- Update documentation
- Fix link formatting in som.py documentation

New dependencies:
- scipy

Input arguments:
- `ci-level` (default: 0.95)

Output:
- New columns
  + 'recall_lower', 'recall_upper' and
  + 'precision_lower', 'precision_upper'

+++ src/sh/run_sompy_test.sh
@@ -226,6 +226,29 @@ fi
 
 #####
 
+echo "Somatic comparison test with custom confidence interval leve for ${HCVERSION} from ${HCDIR}l"
+
+TMP_OUT=`mktemp -t sompy.XXXXXXXXXX`
+
+# run som.py
+${PYTHON} ${HCDIR}/som.py \
+			 	  ${DIR}/../../example/sompy/PG_admix_truth_snvs.vcf.gz \
+			 	  ${DIR}/../../example/sompy/strelka_admix_snvs.vcf.gz \
+			 	  -o ${TMP_OUT} -P --ci-level 0.99
+
+if [[ $? != 0 ]]; then
+	  echo "som.py failed!"
+	  exit 1
+fi
+
+${PYTHON} ${DIR}/compare_sompy.py ${TMP_OUT}.stats.csv ${DIR}/../../example/sompy/stats_ci.csv
+if [[ $? != 0 ]]; then
+	  echo "Output counts differ diff  ${TMP_OUT}.stats.csv ${DIR}/../../example/sompy/stats_ci.csv !"
+	  exit 1
+fi
+
+#####
+
 rm -f ${TMP_OUT}*
 
 echo "Som.py test done"
